Upgrade to mlua 0.11

+++ crates/lune-roblox/src/datatypes/conversion.rs
@@ -88,8 +88,8 @@ impl LuaToDomValue for LuaValue {
             match (self, variant_type) {
                 (LuaValue::Boolean(b), DomType::Bool) => Ok(DomValue::Bool(*b)),
 
-                (LuaValue::Integer(i), DomType::Int64) => Ok(DomValue::Int64(*i as i64)),
-                (LuaValue::Integer(i), DomType::Int32) => Ok(DomValue::Int32(*i)),
+                (LuaValue::Integer(i), DomType::Int64) => Ok(DomValue::Int64(*i)),
+                (LuaValue::Integer(i), DomType::Int32) => Ok(DomValue::Int32(*i as i32)),
                 (LuaValue::Integer(i), DomType::Float64) => Ok(DomValue::Float64(*i as f64)),
                 (LuaValue::Integer(i), DomType::Float32) => Ok(DomValue::Float32(*i as f32)),
 
@@ -126,7 +126,7 @@ impl LuaToDomValue for LuaValue {
         } else {
             match self {
                 LuaValue::Boolean(b) => Ok(DomValue::Bool(*b)),
-                LuaValue::Integer(i) => Ok(DomValue::Int32(*i)),
+                LuaValue::Integer(i) => Ok(DomValue::Int32(*i as i32)),
                 LuaValue::Number(n) => Ok(DomValue::Float64(*n)),
                 LuaValue::String(s) => Ok(DomValue::String(s.to_str()?.to_string())),
                 LuaValue::UserData(u) => u.lua_to_dom_value(lua, None),
